refactor: api auth 폴더구조 정리
